viewの確認は機能テストではないのでフォルダ移動
