+++ tests/View/keiziban/AdminKeizibanNotShowViewTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Tests\Feature\keiziban;
+namespace Tests\View\keiziban;
 
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
 use Tests\TestCase;
 
-class KeizibanShowViewTest extends TestCase
+class AdminKeizibanNotShowViewTest extends TestCase
 {
     /**
      * A basic feature test example.
@@ -15,20 +15,20 @@ class KeizibanShowViewTest extends TestCase
      * @return void
      */
 
-    private $user;
+    private $admin;
 
     public function setUp(): void
     {
         parent::setUp();
-        $this->user = User::factory()->make();
+        $this->admin = User::factory()->admin()->make();
     }
 
     public function test_keiziban_view_have_jQuery()
     {
         $TextView = (string) $this
-            ->actingAs($this->user)
+            ->actingAs($this->admin)
             ->view('keiziban', [
-                'result' => 1,
+                'result' => 0,
                 'thread_name' => 'ThreadTestName',
                 'thread_id' => 'ThreadTestID',
                 'username' => 'TestUser',
@@ -36,18 +36,18 @@ class KeizibanShowViewTest extends TestCase
             ]);
 
         if (strpos($TextView, '<script src="https://code.jquery.com/jquery-3.0.0.min.js"></script>') !== false) {
-            $this->assertAuthenticated($this->user = null);
+            $this->assertAuthenticated($this->admin = null);
         } else {
-            $this->assertGuest($this->user = null);
+            $this->assertGuest($this->admin = null);
         }
     }
 
     public function test_keiziban_view_have_url()
     {
         $TextView = (string) $this
-            ->actingAs($this->user)
+            ->actingAs($this->admin)
             ->view('keiziban', [
-                'result' => 1,
+                'result' => 0,
                 'thread_name' => 'ThreadTestName',
                 'thread_id' => 'ThreadTestID',
                 'username' => 'TestUser',
@@ -57,18 +57,18 @@ class KeizibanShowViewTest extends TestCase
         $text = 'const url = "' . (string) url('/') . '";';
 
         if (strpos($TextView, $text) !== false) {
-            $this->assertAuthenticated($this->user = null);
+            $this->assertAuthenticated($this->admin = null);
         } else {
-            $this->assertGuest($this->user = null);
+            $this->assertGuest($this->admin = null);
         }
     }
 
     public function test_keiziban_view_have_table()
     {
         $TextView = (string) $this
-            ->actingAs($this->user)
+            ->actingAs($this->admin)
             ->view('keiziban', [
-                'result' => 1,
+                'result' => 0,
                 'thread_name' => 'ThreadTestName',
                 'thread_id' => 'ThreadTestID',
                 'username' => 'TestUser',
@@ -78,18 +78,18 @@ class KeizibanShowViewTest extends TestCase
         $text = 'const table = "ThreadTestName";';
 
         if (strpos($TextView, $text) !== false) {
-            $this->assertAuthenticated($this->user = null);
+            $this->assertAuthenticated($this->admin = null);
         } else {
-            $this->assertGuest($this->user = null);
+            $this->assertGuest($this->admin = null);
         }
     }
 
     public function test_keiziban_view_have_thread_id()
     {
         $TextView = (string) $this
-            ->actingAs($this->user)
+            ->actingAs($this->admin)
             ->view('keiziban', [
-                'result' => 1,
+                'result' => 0,
                 'thread_name' => 'ThreadTestName',
                 'thread_id' => 'ThreadTestID',
                 'username' => 'TestUser',
@@ -99,33 +99,33 @@ class KeizibanShowViewTest extends TestCase
         $text = 'const thread_id = "ThreadTestID";';
 
         if (strpos($TextView, $text) !== false) {
-            $this->assertAuthenticated($this->user = null);
+            $this->assertAuthenticated($this->admin = null);
         } else {
-            $this->assertGuest($this->user = null);
+            $this->assertGuest($this->admin = null);
         }
     }
 
-    public function test_keiziban_view_have_likes_function()
+    public function test_keiziban_view_not_have_likes_function()
     {
         $view = $this
-            ->actingAs($this->user)
+            ->actingAs($this->admin)
             ->view('keiziban', [
-                'result' => 1,
+                'result' => 0,
                 'thread_name' => 'ThreadTestName',
                 'thread_id' => 'ThreadTestID',
                 'username' => 'TestUser',
                 'url' => url('/')
             ]);
 
-        $view->assertSee('function likes(message_id, user_like)');
+        $view->assertDontSee('function likes(message_id, user_like)');
     }
 
     public function test_keiziban_view_not_have_keiziban_message_actions_form()
     {
         $view = $this
-            ->actingAs($this->user)
+            ->actingAs($this->admin)
             ->view('keiziban', [
-                'result' => 1,
+                'result' => 0,
                 'thread_name' => 'ThreadTestName',
                 'thread_id' => 'ThreadTestID',
                 'username' => 'TestUser',
@@ -138,9 +138,9 @@ class KeizibanShowViewTest extends TestCase
     public function test_keiziban_view_not_have_keiziban_message_id_number()
     {
         $view = $this
-            ->actingAs($this->user)
+            ->actingAs($this->admin)
             ->view('keiziban', [
-                'result' => 1,
+                'result' => 0,
                 'thread_name' => 'ThreadTestName',
                 'thread_id' => 'ThreadTestID',
                 'username' => 'TestUser',
@@ -150,147 +150,147 @@ class KeizibanShowViewTest extends TestCase
         $view->assertDontSee('keiziban_message_id_number');
     }
 
-    public function test_keiziban_view_have_keiziban_sendMessage_form()
+    public function test_keiziban_view_not_have_keiziban_sendMessage_form()
     {
         $view = $this
-            ->actingAs($this->user)
+            ->actingAs($this->admin)
             ->view('keiziban', [
-                'result' => 1,
+                'result' => 0,
                 'thread_name' => 'ThreadTestName',
                 'thread_id' => 'ThreadTestID',
                 'username' => 'TestUser',
                 'url' => url('/')
             ]);
 
-        $view->assertSee('keiziban_sendMessage_form');
+        $view->assertDontSee('keiziban_sendMessage_form');
     }
 
-    public function test_keiziban_view_have_keiziban_message_textarea()
+    public function test_keiziban_view__not_have_keiziban_message_textarea()
     {
         $view = $this
-            ->actingAs($this->user)
+            ->actingAs($this->admin)
             ->view('keiziban', [
-                'result' => 1,
+                'result' => 0,
                 'thread_name' => 'ThreadTestName',
                 'thread_id' => 'ThreadTestID',
                 'username' => 'TestUser',
                 'url' => url('/')
             ]);
 
-        $view->assertSee('keiziban_message_textarea');
+        $view->assertDontSee('keiziban_message_textarea');
     }
 
-    public function test_keiziban_view_have_keiziban_sendAlertArea()
+    public function test_keiziban_view_not_have_keiziban_sendAlertArea()
     {
         $view = $this
-            ->actingAs($this->user)
+            ->actingAs($this->admin)
             ->view('keiziban', [
-                'result' => 1,
+                'result' => 0,
                 'thread_name' => 'ThreadTestName',
                 'thread_id' => 'ThreadTestID',
                 'username' => 'TestUser',
                 'url' => url('/')
             ]);
 
-        $view->assertSee('keiziban_sendAlertArea');
+        $view->assertDontSee('keiziban_sendAlertArea');
     }
 
-    public function test_keiziban_view_have_keiziban_sendMessage_btn()
+    public function test_keiziban_view_not_have_keiziban_sendMessage_btn()
     {
         $view = $this
-            ->actingAs($this->user)
+            ->actingAs($this->admin)
             ->view('keiziban', [
-                'result' => 1,
+                'result' => 0,
                 'thread_name' => 'ThreadTestName',
                 'thread_id' => 'ThreadTestID',
                 'username' => 'TestUser',
                 'url' => url('/')
             ]);
 
-        $view->assertSee('keiziban_sendMessage_btn');
+        $view->assertDontSee('keiziban_sendMessage_btn');
     }
 
-    public function test_keiziban_view_have_keiziban_displayArea()
+    public function test_keiziban_view_not_have_keiziban_displayArea()
     {
         $view = $this
-            ->actingAs($this->user)
+            ->actingAs($this->admin)
             ->view('keiziban', [
-                'result' => 1,
+                'result' => 0,
                 'thread_name' => 'ThreadTestName',
                 'thread_id' => 'ThreadTestID',
                 'username' => 'TestUser',
                 'url' => url('/')
             ]);
 
-        $view->assertSee('keiziban_displayArea');
+        $view->assertDontSee('keiziban_displayArea');
     }
 
-    public function test_keiziban_view_not_have_keiziban_DeleteMessage_Modal()
+    public function test_keiziban_view_have_keiziban_DeleteMessage_Modal()
     {
         $view = $this
-            ->actingAs($this->user)
+            ->actingAs($this->admin)
             ->view('keiziban', [
-                'result' => 1,
+                'result' => 0,
                 'thread_name' => 'ThreadTestName',
                 'thread_id' => 'ThreadTestID',
                 'username' => 'TestUser',
                 'url' => url('/')
             ]);
 
-        $view->assertDontSee('keiziban_DeleteMessage_Modal');
+        $view->assertSee('keiziban_DeleteMessage_Modal');
     }
 
-    public function test_keiziban_view_not_have_keiziban_delete_message_btn()
+    public function test_keiziban_view_have_keiziban_delete_message_btn()
     {
         $view = $this
-            ->actingAs($this->user)
+            ->actingAs($this->admin)
             ->view('keiziban', [
-                'result' => 1,
+                'result' => 0,
                 'thread_name' => 'ThreadTestName',
                 'thread_id' => 'ThreadTestID',
                 'username' => 'TestUser',
                 'url' => url('/')
             ]);
 
-        $view->assertDontSee('keiziban_delete_message_btn');
+        $view->assertSee('keiziban_delete_message_btn');
     }
 
-    public function test_keiziban_view_not_have_keiziban_RestoreMessage_Modal()
+    public function test_keiziban_view_have_keiziban_RestoreMessage_Modal()
     {
         $view = $this
-            ->actingAs($this->user)
+            ->actingAs($this->admin)
             ->view('keiziban', [
-                'result' => 1,
+                'result' => 0,
                 'thread_name' => 'ThreadTestName',
                 'thread_id' => 'ThreadTestID',
                 'username' => 'TestUser',
                 'url' => url('/')
             ]);
 
-        $view->assertDontSee('keiziban_RestoreMessage_Modal');
+        $view->assertSee('keiziban_RestoreMessage_Modal');
     }
 
-    public function test_keiziban_view_not_have_keiziban_restore_message_btn()
+    public function test_keiziban_view_have_keiziban_restore_message_btn()
     {
         $view = $this
-            ->actingAs($this->user)
+            ->actingAs($this->admin)
             ->view('keiziban', [
-                'result' => 1,
+                'result' => 0,
                 'thread_name' => 'ThreadTestName',
                 'thread_id' => 'ThreadTestID',
                 'username' => 'TestUser',
                 'url' => url('/')
             ]);
 
-        $view->assertDontSee('keiziban_restore_message_btn');
+        $view->assertSee('keiziban_restore_message_btn');
     }
 
     public function test_keiziban_view_have_js()
     {
         $TextView = (string) $this
-            ->actingAs($this->user)
+            ->actingAs($this->admin)
             ->view('keiziban', [
-                'result' => 1,
+                'result' => 0,
                 'thread_name' => 'ThreadTestName',
                 'thread_id' => 'ThreadTestID',
                 'username' => 'TestUser',
@@ -300,9 +300,9 @@ class KeizibanShowViewTest extends TestCase
         $text = `<script src="{{ mix('js/app_jquery.js') }}"></script>`;
 
         if (strpos($TextView, $text) !== false) {
-            $this->assertAuthenticated($this->user = null);
+            $this->assertAuthenticated($this->admin = null);
         } else {
-            $this->assertGuest($this->user = null);
+            $this->assertGuest($this->admin = null);
         }
     }
 }
