+++ tests/View/keiziban/KeizibanShowViewTest.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Tests\Feature\keiziban;
+namespace Tests\View\keiziban;
 
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
 use Tests\TestCase;
 
-class AdminKeizibanShowViewTest extends TestCase
+class KeizibanShowViewTest extends TestCase
 {
     /**
      * A basic feature test example.
@@ -15,18 +15,18 @@ class AdminKeizibanShowViewTest extends TestCase
      * @return void
      */
 
-    private $admin;
+    private $user;
 
     public function setUp(): void
     {
         parent::setUp();
-        $this->admin = User::factory()->admin()->make();
+        $this->user = User::factory()->make();
     }
 
     public function test_keiziban_view_have_jQuery()
     {
         $TextView = (string) $this
-            ->actingAs($this->admin)
+            ->actingAs($this->user)
             ->view('keiziban', [
                 'result' => 1,
                 'thread_name' => 'ThreadTestName',
@@ -36,16 +36,16 @@ class AdminKeizibanShowViewTest extends TestCase
             ]);
 
         if (strpos($TextView, '<script src="https://code.jquery.com/jquery-3.0.0.min.js"></script>') !== false) {
-            $this->assertAuthenticated($this->admin = null);
+            $this->assertAuthenticated($this->user = null);
         } else {
-            $this->assertGuest($this->admin = null);
+            $this->assertGuest($this->user = null);
         }
     }
 
     public function test_keiziban_view_have_url()
     {
         $TextView = (string) $this
-            ->actingAs($this->admin)
+            ->actingAs($this->user)
             ->view('keiziban', [
                 'result' => 1,
                 'thread_name' => 'ThreadTestName',
@@ -57,16 +57,16 @@ class AdminKeizibanShowViewTest extends TestCase
         $text = 'const url = "' . (string) url('/') . '";';
 
         if (strpos($TextView, $text) !== false) {
-            $this->assertAuthenticated($this->admin = null);
+            $this->assertAuthenticated($this->user = null);
         } else {
-            $this->assertGuest($this->admin = null);
+            $this->assertGuest($this->user = null);
         }
     }
 
     public function test_keiziban_view_have_table()
     {
         $TextView = (string) $this
-            ->actingAs($this->admin)
+            ->actingAs($this->user)
             ->view('keiziban', [
                 'result' => 1,
                 'thread_name' => 'ThreadTestName',
@@ -78,16 +78,16 @@ class AdminKeizibanShowViewTest extends TestCase
         $text = 'const table = "ThreadTestName";';
 
         if (strpos($TextView, $text) !== false) {
-            $this->assertAuthenticated($this->admin = null);
+            $this->assertAuthenticated($this->user = null);
         } else {
-            $this->assertGuest($this->admin = null);
+            $this->assertGuest($this->user = null);
         }
     }
 
     public function test_keiziban_view_have_thread_id()
     {
         $TextView = (string) $this
-            ->actingAs($this->admin)
+            ->actingAs($this->user)
             ->view('keiziban', [
                 'result' => 1,
                 'thread_name' => 'ThreadTestName',
@@ -99,16 +99,16 @@ class AdminKeizibanShowViewTest extends TestCase
         $text = 'const thread_id = "ThreadTestID";';
 
         if (strpos($TextView, $text) !== false) {
-            $this->assertAuthenticated($this->admin = null);
+            $this->assertAuthenticated($this->user = null);
         } else {
-            $this->assertGuest($this->admin = null);
+            $this->assertGuest($this->user = null);
         }
     }
 
     public function test_keiziban_view_have_likes_function()
     {
         $view = $this
-            ->actingAs($this->admin)
+            ->actingAs($this->user)
             ->view('keiziban', [
                 'result' => 1,
                 'thread_name' => 'ThreadTestName',
@@ -120,10 +120,10 @@ class AdminKeizibanShowViewTest extends TestCase
         $view->assertSee('function likes(message_id, user_like)');
     }
 
-    public function test_keiziban_view_have_keiziban_message_actions_form()
+    public function test_keiziban_view_not_have_keiziban_message_actions_form()
     {
         $view = $this
-            ->actingAs($this->admin)
+            ->actingAs($this->user)
             ->view('keiziban', [
                 'result' => 1,
                 'thread_name' => 'ThreadTestName',
@@ -132,13 +132,13 @@ class AdminKeizibanShowViewTest extends TestCase
                 'url' => url('/')
             ]);
 
-        $view->assertSee('keiziban_message_actions_form');
+        $view->assertDontSee('keiziban_message_actions_form');
     }
 
-    public function test_keiziban_view_have_keiziban_message_id_number()
+    public function test_keiziban_view_not_have_keiziban_message_id_number()
     {
         $view = $this
-            ->actingAs($this->admin)
+            ->actingAs($this->user)
             ->view('keiziban', [
                 'result' => 1,
                 'thread_name' => 'ThreadTestName',
@@ -147,13 +147,13 @@ class AdminKeizibanShowViewTest extends TestCase
                 'url' => url('/')
             ]);
 
-        $view->assertSee('keiziban_message_id_number');
+        $view->assertDontSee('keiziban_message_id_number');
     }
 
     public function test_keiziban_view_have_keiziban_sendMessage_form()
     {
         $view = $this
-            ->actingAs($this->admin)
+            ->actingAs($this->user)
             ->view('keiziban', [
                 'result' => 1,
                 'thread_name' => 'ThreadTestName',
@@ -168,7 +168,7 @@ class AdminKeizibanShowViewTest extends TestCase
     public function test_keiziban_view_have_keiziban_message_textarea()
     {
         $view = $this
-            ->actingAs($this->admin)
+            ->actingAs($this->user)
             ->view('keiziban', [
                 'result' => 1,
                 'thread_name' => 'ThreadTestName',
@@ -183,7 +183,7 @@ class AdminKeizibanShowViewTest extends TestCase
     public function test_keiziban_view_have_keiziban_sendAlertArea()
     {
         $view = $this
-            ->actingAs($this->admin)
+            ->actingAs($this->user)
             ->view('keiziban', [
                 'result' => 1,
                 'thread_name' => 'ThreadTestName',
@@ -198,7 +198,7 @@ class AdminKeizibanShowViewTest extends TestCase
     public function test_keiziban_view_have_keiziban_sendMessage_btn()
     {
         $view = $this
-            ->actingAs($this->admin)
+            ->actingAs($this->user)
             ->view('keiziban', [
                 'result' => 1,
                 'thread_name' => 'ThreadTestName',
@@ -213,7 +213,7 @@ class AdminKeizibanShowViewTest extends TestCase
     public function test_keiziban_view_have_keiziban_displayArea()
     {
         $view = $this
-            ->actingAs($this->admin)
+            ->actingAs($this->user)
             ->view('keiziban', [
                 'result' => 1,
                 'thread_name' => 'ThreadTestName',
@@ -225,10 +225,10 @@ class AdminKeizibanShowViewTest extends TestCase
         $view->assertSee('keiziban_displayArea');
     }
 
-    public function test_keiziban_view_have_keiziban_DeleteMessage_Modal()
+    public function test_keiziban_view_not_have_keiziban_DeleteMessage_Modal()
     {
         $view = $this
-            ->actingAs($this->admin)
+            ->actingAs($this->user)
             ->view('keiziban', [
                 'result' => 1,
                 'thread_name' => 'ThreadTestName',
@@ -237,13 +237,13 @@ class AdminKeizibanShowViewTest extends TestCase
                 'url' => url('/')
             ]);
 
-        $view->assertSee('keiziban_DeleteMessage_Modal');
+        $view->assertDontSee('keiziban_DeleteMessage_Modal');
     }
 
-    public function test_keiziban_view_have_keiziban_delete_message_btn()
+    public function test_keiziban_view_not_have_keiziban_delete_message_btn()
     {
         $view = $this
-            ->actingAs($this->admin)
+            ->actingAs($this->user)
             ->view('keiziban', [
                 'result' => 1,
                 'thread_name' => 'ThreadTestName',
@@ -252,13 +252,13 @@ class AdminKeizibanShowViewTest extends TestCase
                 'url' => url('/')
             ]);
 
-        $view->assertSee('keiziban_delete_message_btn');
+        $view->assertDontSee('keiziban_delete_message_btn');
     }
 
-    public function test_keiziban_view_have_keiziban_RestoreMessage_Modal()
+    public function test_keiziban_view_not_have_keiziban_RestoreMessage_Modal()
     {
         $view = $this
-            ->actingAs($this->admin)
+            ->actingAs($this->user)
             ->view('keiziban', [
                 'result' => 1,
                 'thread_name' => 'ThreadTestName',
@@ -267,13 +267,13 @@ class AdminKeizibanShowViewTest extends TestCase
                 'url' => url('/')
             ]);
 
-        $view->assertSee('keiziban_RestoreMessage_Modal');
+        $view->assertDontSee('keiziban_RestoreMessage_Modal');
     }
 
-    public function test_keiziban_view_have_keiziban_restore_message_btn()
+    public function test_keiziban_view_not_have_keiziban_restore_message_btn()
     {
         $view = $this
-            ->actingAs($this->admin)
+            ->actingAs($this->user)
             ->view('keiziban', [
                 'result' => 1,
                 'thread_name' => 'ThreadTestName',
@@ -282,13 +282,13 @@ class AdminKeizibanShowViewTest extends TestCase
                 'url' => url('/')
             ]);
 
-        $view->assertSee('keiziban_restore_message_btn');
+        $view->assertDontSee('keiziban_restore_message_btn');
     }
 
     public function test_keiziban_view_have_js()
     {
         $TextView = (string) $this
-            ->actingAs($this->admin)
+            ->actingAs($this->user)
             ->view('keiziban', [
                 'result' => 1,
                 'thread_name' => 'ThreadTestName',
@@ -300,9 +300,9 @@ class AdminKeizibanShowViewTest extends TestCase
         $text = `<script src="{{ mix('js/app_jquery.js') }}"></script>`;
 
         if (strpos($TextView, $text) !== false) {
-            $this->assertAuthenticated($this->admin = null);
+            $this->assertAuthenticated($this->user = null);
         } else {
-            $this->assertGuest($this->admin = null);
+            $this->assertGuest($this->user = null);
         }
     }
 }
